+++ plugin/pkg/kubernetes/object/informer.go
